requirements.txt
redis
requests
boto3
python-dotenv
sseclient-py
colorama